Refactor `compose.py` module to `compose` package with `_func_fitter.py` and `_pipeline.py` modules
